put all admin page actions in a separate file
